New SkyblockItem.kt and InventoryUtils.kt utility classes.
The use of item names was replaced in many places by the new SkyblockItem.kt Enum.
When the enum is used the item id is used instead of the item name, which avoids conflicts with custom item names from other mods.

Utility functions were expanded to also use the new enum which will also make it easier to make new modules and fixing things when new items are added to the game.