GUI Rework #2
BIG code cleanup.
The structure of all the classes is now a lot more consistent and the code a lot simpler.

Rendering and interactions are all dispatched by the ClickGUI to the top level components of the gui (the panels).
The components of the gui dispatch all actions and rendering to their subcomponents.

Improved documentation

New methods for HUDRenderUtils to help with setting up scissor tests.