Big backend changes which make creating new modules easier. (#11)
* New DummySetting Class.
It is used as a default instance of Setting without any new functionality
This allows for making the Setting Class  abstract.

* Setting Class Rework #1
centralized value, default and process input properties.
Implemented operator methods which enabled the use of Settings as a delegate.
New DummySetting class that can be used as a generic setting to avoid using a nullable type.

* Module Class Rework #1
Improved documentation.
New ConfigModule dummy class. This allows for Module to be abstract.

* Simplified the ModuleConfig class.

* Added full delegation support for the Setting class inside of module classes.
Also improved documentation

* Reworked SelectorSetting to work with enums instead of Strings.

Kept support for the old version and marked everything related as deprecated.

* Improved doc comments for the Setting classes.

* GUI Rework #2

BIG code cleanup.
The structure of all the classes is now a lot more consistent and the code a lot simpler.

Rendering and interactions are all dispatched by the ClickGUI to the top level components of the gui (the panels).
The components of the gui dispatch all actions and rendering to their subcomponents.

Improved documentation

New methods for HUDRenderUtils to help with setting up scissor tests.

* HUD element rework #1

New Annotation RegisterHudElement makes the module manager automatically register the hud element to the module and to the edit hud gui.

* HUD element rework #2

Added a new constructor to the HudElement class, which makes creating hud elemets a lot more streamlined.
It creates the settings for you.

* Module rework #2

Added a new annotation SelfRegisterModule for Modules.
This annotation will make the module manager automatically load in the module without the need for it to be in the ModuleManager.modules list.